pyproject.toml
[project]
name = "convflow-markdown"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "markitdown[pptx,docx,xlsx,xls,pdf,outlook]>=0.1.2",
    "fastapi>=0.104.0",
    "uvicorn>=0.24.0",
    "python-multipart>=0.0.6",
    "httpx>=0.25.0",
    "python-dotenv>=1.0.0",
    "pyjwt>=2.8.0",
    "bcrypt>=4.0.0",
    "asyncpg>=0.29.0",
    "pydantic[email]>=2.0.0",
]
